move copilot upstream discovery to internal
